[Refactor] : eb-server -> eb_fast_api 이름변경
Git : #8
