+++ api/eb_fast_api/database/database.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-DB_URL = "sqlite:///./eb_server/database/earlybuddy.db"
+DB_URL = "sqlite:///./eb_fast_api/database/earlybuddy.db"
 
 engine = create_engine(DB_URL, connect_args={"check_same_thread" : False})
 SessionLocal = sessionmaker(autoflush=False, bind=engine)
