+++ api/requirements.txt
@@ -7,7 +7,7 @@ charset-normalizer==3.3.2
 click==8.1.7
 cryptography==42.0.5
 dnspython==2.6.1
-eb_server==0.1.0
+eb_fast_api==0.1.0
 ecdsa==0.19.0
 email_validator==2.1.1
 fastapi==0.110.0
