--- eb_server.Dockerfile
@@ -1,7 +1,0 @@
-FROM python:3.12.3-slim
-
-WORKDIR /mycode
-COPY . .
-
-RUN pip install -e .
-RUN pip install -r requirements.txt
